entg: introduce two build modes (host/occlum) for entg
This commits introduced two build target for entg: entg-occlum and
entg-host. They are different binary file, in path
`target/debug/entg-occlum` and `target/debug/entg-host`. At this moment,
the difference between them is the different librats_tls.so file they
will link to (see entg/build.rs).

We utilize the source code in deps/rats-tls to build librats_tls.so
separately for occlum and host mode. For the consideration of avoiding
conflicts, we only run `make install` for host mode, the libraries of
occlum mode will not be installed into the system. For more information,
please check the Makefile