rats-tls: fix meaningless waiting when process exiting
The problem is caused by two reasons. First, the reader task of rats-tls
did not shutdown the write-half of DuplexStream, which cause the main
thread to keep waiting to read from the DuplexStream. Second, the writer
task of rats-tls did shutdown the TcpStream before exiting, which cause
the meaningless waiting on the peer of the rats-tls connection.